requirements.txt
astroid==2.2.5
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==2.5.3
django-debug-toolbar==1.11
django-filter==2.1.0
django-pg-commands==0.1.3
djangorestframework==3.9.3
djangorestframework-recursive==0.1.2
idna==2.8
include==0.2.2
inquirer==2.5.1
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pylint==2.3.1
pytz==2019.1
readchar==2.0.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.2
wrapt==1.11.1
